+++ example/lib/main.dart
@@ -21,8 +21,21 @@ class Example extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       body: PanelsTheme(
-        data: PanelsThemeData(), // FrostedPanelsThemeData()
-        child: PanelsManager(children: [MyCustomWidget()]),
+        data: FrostedPanelsThemeData(), // FrostedPanelsThemeData() // PanelsThemeData()
+        child: PanelsManager(
+          children: [MyCustomWidget()],
+          initialPanels: [],
+          childrenOnTop: [
+            Row(
+              children: [
+                Material(
+                  color: Theme.of(context).primaryColor,
+                  child: Text("I'm on top of all panels!!!"),
+                )
+              ],
+            )
+          ],
+        ),
       ),
     );
   }
@@ -30,12 +43,15 @@ class Example extends StatelessWidget {
 
 
 class MyCustomWidget extends StatefulWidget {
+  // GlobalKey is used here in order to preserve state when moved around in the widget tree
+  MyCustomWidget() : super(key: GlobalKey());
+
   @override
   _MyCustomWidgetState createState() => _MyCustomWidgetState();
 }
 
 class _MyCustomWidgetState extends State<MyCustomWidget> {
-  int windowsOpened = 0;
+  int panelsOpened = 0;
 
   @override
   Widget build(BuildContext context) {
@@ -45,12 +61,12 @@ class _MyCustomWidgetState extends State<MyCustomWidget> {
         children: [
           Expanded(
             child: Center(
-              child: Text("My Custom Widget"),
+              child: Text("My Custom Widget", textAlign: TextAlign.center,),
             ),
           ),
 
           Expanded(
-            child: Text("This widget has opened $windowsOpened windows."),
+            child: Text("This widget has opened $panelsOpened panels.", textAlign: TextAlign.center),
           ),
 
           Expanded(
@@ -58,12 +74,21 @@ class _MyCustomWidgetState extends State<MyCustomWidget> {
               child: MaterialButton(
                 color: Theme.of(context).primaryColor,
                 onPressed: () {
-                  Panels.of(context).addPanel(widget: MyCustomWidget(), title: "Opened From MyCustomWidget()");
+                  Panels.of(context).addPanel(
+                    Panel(
+                      initialTabs: [
+                        PanelTab(
+                          child: MyCustomWidget(),
+                          title: "Opened From MyCustomWidget()"
+                        )
+                      ],
+                    )
+                  );
                   setState(() {
-                    windowsOpened++;
+                    panelsOpened++;
                   });
                 },
-                child: Text("Open New Window"),
+                child: Text("Open New Panel", textAlign: TextAlign.center),
               ),
             ),
           )